+++ pyembed/src/test/test_importer_module.py
@@ -8,13 +8,13 @@ import unittest
 
 class TestImporterModule(unittest.TestCase):
     def test_module(self):
-        import _pyoxidizer_importer as importer
+        import oxidized_importer as importer
 
         attrs = {a for a in dir(importer) if not a.startswith("__")}
         self.assertEqual(attrs, {"decode_source", "OxidizedResource", "OxidizedFinder"})
 
     def test_finder_attrs(self):
-        from _pyoxidizer_importer import OxidizedFinder
+        from oxidized_importer import OxidizedFinder
 
         attrs = {a for a in dir(OxidizedFinder) if not a.startswith("__")}
         self.assertEqual(